# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SeamLabs APIs",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/problems/first": {
            "post": {
                "tags": [
                    "problems"
                ],
                "summary": "First problem",
                "description": "First problem",
                "operationId": "ab21a7ad834df03158515935e35128d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                }
                            },
                            "example": "[5,-9,-8,4,-9,4,7,10,7]"
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns the smallest missing integer"
                    },
                    "422": {
                        "description": "Validation error/missing parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The provided parameters are incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/problems/second/{start}/{end}": {
            "get": {
                "tags": [
                    "problems"
                ],
                "summary": "Second problem",
                "description": "Second problem",
                "operationId": "f6533d92673a8f8118edc3db93db341d",
                "parameters": [
                    {
                        "name": "start",
                        "in": "path",
                        "description": "start integer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "end",
                        "in": "path",
                        "description": "end integer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns count of all number except numbers that include the digit 5"
                    },
                    "422": {
                        "description": "Validation error/missing parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The provided parameters are incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/problems/third/{str}": {
            "get": {
                "tags": [
                    "problems"
                ],
                "summary": "Third problem",
                "description": "Third problem",
                "operationId": "5db98e73e024c6a1021a544b9de581f8",
                "parameters": [
                    {
                        "name": "str",
                        "in": "path",
                        "description": "alphabetic string",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "BFG"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the alphabetic index of the string"
                    },
                    "422": {
                        "description": "Validation error/missing parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The provided parameters are incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/problems/fourth": {
            "post": {
                "tags": [
                    "problems"
                ],
                "summary": "Fourth problem",
                "description": "Fourth problem",
                "operationId": "86289dd4dadf6d708fe2eb2071e484dd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "N": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "Q": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Validation error/missing parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The provided parameters are incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}